A few refactors for parent nodes abstraction
